+++ R/get_targets.R
@@ -10,13 +10,13 @@
 ##' 'STITCH' or 'all'. If 'all', the targets from DrugBank, CLUE 
 ##' and STITCH databases will be combined.  
 ##' @param verbose TRUE or FALSE, whether to print messages
-##' @param collapse TRUE or FALSE. If setting as FALSE, the result is a 
-##' data.frame of target gene symbols of each drug. If TRUE, the targets are 
-##' collapsed to be a character vector with duplications if different drugs 
+##' @param output one of "df", "list" or "vector". If setting as "df", the result 
+##' is in a data.frame format containing target gene symbols separated by semicolon
+##' for each drug. If setting as "list", the result is a list of targets for each
+##' query drug. If setting as "vector", the result is a character vector of the 
+##' target set that are collapsed with duplications if different drugs 
 ##' have the same targets.
-##' @return data.frame, one column contains the query drug names and the other 
-##' target gene symbols if \code{collapse} is FALSE. A character vetor if
-##' \code{collapse} is TRUE.
+##' @return drug-target annotation in a format defined by the \code{output} argument.
 ##' @importFrom RSQLite dbConnect
 ##' @importFrom RSQLite dbGetQuery
 ##' @importFrom RSQLite SQLite
@@ -27,7 +27,7 @@
 ##' dt <- get_targets(drugs10)
 ##' @export 
 
-get_targets <- function(drugs, database="all", verbose=TRUE, collapse=FALSE){
+get_targets <- function(drugs, database="all", verbose=TRUE, output="df"){
   drugs_orig <- unique(drugs)
   drugs <- unique(tolower(drugs))
   # load dtlink_db_clue_sti.db stored in AnnotationHub
@@ -38,77 +38,39 @@ get_targets <- function(drugs, database="all", verbose=TRUE, collapse=FALSE){
   dtlink <- dbGetQuery(conn, 'SELECT * FROM dtlink')
   dbDisconnect(conn)
   
-  # drug targets in DrugBank
-  if(database=="DrugBank"){
-      dtlist_db <- split(dtlink_db$t_gn_sym, dtlink_db$drug_name)
-      drugs_notar_db <- setdiff(drugs, names(dtlist_db))
-      dtlist_db_drugs <- dtlist_db[intersect(drugs, names(dtlist_db))]
-      res <- list2slash(dtlist_db_drugs)
-      idx_db <- match(res$drug_name, drugs)
-      res$drug_name <- drugs_orig[idx_db]
-      if(length(drugs_notar_db) > 0 & verbose){
-          message("No targets found in DrugBank database for ", 
-                  length(drugs_notar_db),
-                  " drugs: \n",
-                  paste(drugs_notar_db, collapse = " / "))
+  dtl <- lapply(database, function(x){
+      if(tolower(x)=="drugbank"){
+          return(dtlink_db)
       }
-  }
-  # drug targets in CLUE
-  if(database=="CLUE"){
-      dtlist_clue <- split(dtlink_clue$t_gn_sym, dtlink_clue$drug_name)
-      drugs_notar_clue <- setdiff(drugs, names(dtlist_clue))
-      dtlist_clue_drugs <- dtlist_clue[intersect(drugs, names(dtlist_clue))]
-      res <- list2slash(dtlist_clue_drugs)
-      idx_clue <- match(res$drug_name, drugs)
-      res$drug_name <- drugs_orig[idx_clue]
-      
-      if(length(drugs_notar_clue) > 0 & verbose){
-          message("No targets found in CLUE database for ", 
-                  length(drugs_notar_clue), 
-                  " drugs: \n",
-                  paste(drugs_notar_clue, collapse = " / "))
+      if(tolower(x)=="stitch"){
+          return(dtlink_sti)
       }
-  }
+      if(tolower(x)=="clue"){
+          return(dtlink_clue)
+      }
+      if(tolower(x)=="all"){
+          return(dtlink)
+      }
+  })
+  dt <- unique(do.call(rbind, dtl))
   
-  # drug targets in STITCH
-  if(database=="STITCH"){
-      dtlist_sti <- split(dtlink_sti$t_gn_sym, dtlink_sti$drug_name)
-      drugs_notar_sti <- setdiff(drugs, names(dtlist_sti))
-      dtlist_sti_drugs <- dtlist_sti[intersect(drugs, names(dtlist_sti))]
-      res <- list2slash(dtlist_sti_drugs)
-      idx_sti <- match(res$drug_name, drugs)
-      res$drug_name <- drugs_orig[idx_sti]
-    
-      if(length(drugs_notar_sti) > 0 & verbose){
-          message("No targets found in STITCH database for ", 
-                  length(drugs_notar_sti), 
-                  " drugs: \n",
-                  paste(drugs_notar_sti, collapse = " / "))
-      }
+  dtlist <- split(dt$t_gn_sym, dt$drug_name)
+  drugs_notar <- setdiff(drugs, names(dtlist))
+  if(length(drugs_notar) > 0 & verbose){
+      message("No targets found in ", paste(database, collapse="/"), 
+              " databases for ", length(drugs_notar), " drugs: \n",
+              paste(drugs_notar, collapse = " / "), "\n")
   }
-  
-  # drug targets in all database
-  if(database=="all"){
-      dtlist <- split(dtlink$t_gn_sym, dtlink$drug_name)
-      drugs_notar <- setdiff(drugs, names(dtlist))
-      dtlist_drugs <- dtlist[intersect(drugs, names(dtlist))]
+  dtlist_drugs <- dtlist[intersect(drugs, names(dtlist))]
+  names(dtlist_drugs) <- drugs_orig[match(names(dtlist_drugs), drugs)]
+  if(output=="list"){
+      return(dtlist_drugs)
+  }
+  if(output=="vector"){
+      return(unlist(dtlist_drugs, use.names=FALSE))
+  }
+  if(output=="df"){
       res <- list2slash(dtlist_drugs)
-      idx <- match(res$drug_name, drugs)
-      res$drug_name <- drugs_orig[idx]
-  
-      if(length(drugs_notar) > 0 & verbose){
-          message("No targets found in DrugBank/CLUE/STITCH database for ", 
-                  length(drugs_notar), 
-                  " drugs: \n",
-                  paste(drugs_notar, collapse = " / "), "\n")
-      }
-  }
-  
-  if(collapse){
-      gnset <- na.omit(unlist(lapply(res$t_gn_sym, function(i) 
-          unlist(strsplit(as.character(i), split = "; ")))))
-      return(gnset)
-  } else {
       return(res)
   }
 }