+++ NEWS.md
@@ -1,3 +1,9 @@
+### Changes in version 1.7.3 (2021-08-24)
++ Improved `get_targets` function by supporting different output format.
+
+### Changes in version 1.7.2 (2021-06-14)
++ Supported automatic downloads of ExperimentHub cached files.
+
 ### Changes in version 1.5.8 (2021-05-18)
 + Added FDA phase and other information to `lincs_pert_info` from ChEMBL db.